Use float type for rating to get a more exact prediction
The IMDb rating is a decimal score. Using that may yield more accurate predictions.

+++ app/models/rating.rb
@@ -6,5 +6,5 @@ class Rating
   type 'RATED'
   creates_unique_rel
 
-  property :score, type: Integer
+  property :score, type: Float
 end
